selsa-backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
crispy-bootstrap5==2025.6
cryptography==44.0.3
defusedxml==0.7.1
Django==4.2.21
django-admin-sortable2==2.2.8
django-cas-ng==5.0.1
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-extensions==4.1
django-filter==25.1
django-formtools==2.5.1
django-invitations==2.1.0
django-js-asset==3.1.2
django-mptt==0.17.0
django-otp==1.6.0
django-phonenumber-field==8.1.0
django-two-factor-auth==1.17.0
djangoql==0.18.1
djangorestframework==3.16.0
idna==3.10
lxml==5.4.0
oauthlib==3.2.2
phonenumbers==9.0.7
pillow==11.2.1
ply==3.11
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pypng==0.20220715.0
python-cas==1.6.0
python-decouple==3.8
python3-openid==3.2.0
qrcode==7.4.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
stripe==12.2.0
typing_extensions==4.13.2
urllib3==2.4.0
